requirements.txt
pytorch-lightning
torch~=2.3.0
numpy
pandas
PyYAML~=6.0.1
munch~=4.0.0
wandb~=0.16.6